pyproject.toml
[tool.poetry]
name = "scrapewebdriver"
version = "1.0"
description = "Scrapy extension for handle pages with selenium"
authors = ["aleksey <kvyatkovsky89@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.29.0"
scrapy = "^2.12.0"
beautifulsoup4 = "^4.13.3"
webdriver-manager = "^4.0.2"
pyvirtualdisplay = "^3.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-twisted = "^1.14.3"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
